[refactor] custom hook 제작 error #775
