+++ components/error/Error.tsx
@@ -1,21 +1,9 @@
-import { useRouter } from 'next/router';
-import React, { useEffect } from 'react';
-import { useRecoilState } from 'recoil';
-import { errorState } from 'utils/recoil/error';
+import React from 'react';
+import { useErrorPage } from 'hooks/error/useErrorPage';
 import styles from 'styles/Error.module.scss';
 
 export default function ErrorPage() {
-  const [error, setError] = useRecoilState(errorState);
-  const router = useRouter();
-
-  useEffect(() => {
-    router.replace(`/`);
-  }, []);
-
-  const goHome = () => {
-    setError('');
-    router.push('/');
-  };
+  const { error, goHome } = useErrorPage();
 
   return (
     <div className={styles.container}>
